refactor: Update lexer transformation names and introduce new blank line handling
- Renamed transformation functions for whitespace and indentation processing to `normalize_whitespace` and `sem_indentation`.
- Added a new module for blank line transformations, which converts consecutive newline tokens into semantic blank line tokens.
- Updated references in the lexer and detokenizer to reflect the new function names and modules.
- Ensured that the new transformations preserve source locations correctly.
